requirements.txt
sqlalchemy
pymysql
flask
flasgger
python-dotenv